enhancement: add platform information to analytics user identity
